Move read::* code to the place where it belongs
