+++ libimagtodo/src/lib.rs
@@ -8,7 +8,6 @@ extern crate task_hookrs;
 module_entry_path_mod!("todo", "0.1.0");
 
 pub mod error;
-pub mod read;
 pub mod result;
 pub mod task;
